requirements.txt
# Track-V Backend Requirements

# Framework
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
supabase==1.0.3
psycopg2-binary==2.9.6
sqlalchemy==2.0.0

# Video & Computer Vision
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72
ultralytics==8.0.0
numpy==1.24.3
pillow==10.0.0

# Video Streaming & Download
yt-dlp==2023.3.4
requests==2.31.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
werkzeug==2.3.0

# Email & Notifications
secure==0.3.0

# Production Server
gunicorn==20.1.0
